+++ pkg/apis/kubeovn/v1/register.go
@@ -57,6 +57,10 @@ func addKnownTypes(scheme *runtime.Scheme) error {
 		&SecurityGroupList{},
 		&HtbQos{},
 		&HtbQosList{},
+		&SwitchLBRule{},
+		&SwitchLBRuleList{},
+		&VpcDns{},
+		&VpcDnsList{},
 	)
 	metav1.AddToGroupVersion(scheme, SchemeGroupVersion)
 	return nil