+++ pkg/client/clientset/versioned/typed/kubeovn/v1/kubeovn_client.go
@@ -37,9 +37,11 @@ type KubeovnV1Interface interface {
 	ProviderNetworksGetter
 	SecurityGroupsGetter
 	SubnetsGetter
+	SwitchLBRulesGetter
 	VipsGetter
 	VlansGetter
 	VpcsGetter
+	VpcDnsesGetter
 	VpcNatGatewaysGetter
 }
 
@@ -84,6 +86,10 @@ func (c *KubeovnV1Client) Subnets() SubnetInterface {
 	return newSubnets(c)
 }
 
+func (c *KubeovnV1Client) SwitchLBRules() SwitchLBRuleInterface {
+	return newSwitchLBRules(c)
+}
+
 func (c *KubeovnV1Client) Vips() VipInterface {
 	return newVips(c)
 }
@@ -94,6 +100,10 @@ func (c *KubeovnV1Client) Vlans() VlanInterface {
 
 func (c *KubeovnV1Client) Vpcs() VpcInterface {
 	return newVpcs(c)
+}
+
+func (c *KubeovnV1Client) VpcDnses() VpcDnsInterface {
+	return newVpcDnses(c)
 }
 
 func (c *KubeovnV1Client) VpcNatGateways() VpcNatGatewayInterface {