+++ pkg/client/informers/externalversions/kubeovn/v1/interface.go
@@ -42,12 +42,16 @@ type Interface interface {
 	SecurityGroups() SecurityGroupInformer
 	// Subnets returns a SubnetInformer.
 	Subnets() SubnetInformer
+	// SwitchLBRules returns a SwitchLBRuleInformer.
+	SwitchLBRules() SwitchLBRuleInformer
 	// Vips returns a VipInformer.
 	Vips() VipInformer
 	// Vlans returns a VlanInformer.
 	Vlans() VlanInformer
 	// Vpcs returns a VpcInformer.
 	Vpcs() VpcInformer
+	// VpcDnses returns a VpcDnsInformer.
+	VpcDnses() VpcDnsInformer
 	// VpcNatGateways returns a VpcNatGatewayInformer.
 	VpcNatGateways() VpcNatGatewayInformer
 }
@@ -108,6 +112,11 @@ func (v *version) Subnets() SubnetInformer {
 	return &subnetInformer{factory: v.factory, tweakListOptions: v.tweakListOptions}
 }
 
+// SwitchLBRules returns a SwitchLBRuleInformer.
+func (v *version) SwitchLBRules() SwitchLBRuleInformer {
+	return &switchLBRuleInformer{factory: v.factory, tweakListOptions: v.tweakListOptions}
+}
+
 // Vips returns a VipInformer.
 func (v *version) Vips() VipInformer {
 	return &vipInformer{factory: v.factory, tweakListOptions: v.tweakListOptions}
@@ -121,6 +130,11 @@ func (v *version) Vlans() VlanInformer {
 // Vpcs returns a VpcInformer.
 func (v *version) Vpcs() VpcInformer {
 	return &vpcInformer{factory: v.factory, tweakListOptions: v.tweakListOptions}
+}
+
+// VpcDnses returns a VpcDnsInformer.
+func (v *version) VpcDnses() VpcDnsInformer {
+	return &vpcDnsInformer{factory: v.factory, tweakListOptions: v.tweakListOptions}
 }
 
 // VpcNatGateways returns a VpcNatGatewayInformer.