+++ pkg/controller/pod.go
@@ -685,6 +685,12 @@ func (c *Controller) handleDeletePod(pod *v1.Pod) error {
 		if err = c.deleteAttachmentNetWorkIP(pod); err != nil {
 			klog.Errorf("failed to delete attach ip for pod %v, %v, please delete attach ip manually", pod.Name, err)
 		}
+		if pod.Annotations[util.VipAnnotation] != "" {
+			if err = c.releaseVip(pod.Annotations[util.VipAnnotation]); err != nil {
+				klog.Errorf("failed to clean label from vip %s, %v", pod.Annotations[util.VipAnnotation], err)
+				return err
+			}
+		}
 	}
 	c.ipam.ReleaseAddressByPod(key)
 
@@ -1191,6 +1197,22 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 	podName := c.getNameByPod(pod)
 	key := fmt.Sprintf("%s/%s", pod.Namespace, podName)
 
+	isStsPod, _ := isStatefulSetPod(pod)
+	// if pod has static vip
+	vipName := pod.Annotations[util.VipAnnotation]
+	if vipName != "" {
+		vip, err := c.virtualIpsLister.Get(vipName)
+		if err != nil {
+			klog.Errorf("failed to get static vip '%s', %v", vipName, err)
+			return "", "", "", podNet.Subnet, err
+		}
+		portName := ovs.PodNameToPortName(podName, pod.Namespace, podNet.ProviderName)
+		if err = c.podReuseVip(vipName, portName, isStsPod); err != nil {
+			return "", "", "", podNet.Subnet, err
+		}
+		return vip.Spec.V4ip, vip.Spec.V6ip, vip.Spec.MacAddress, podNet.Subnet, nil
+	}
+
 	macStr := pod.Annotations[fmt.Sprintf(util.MacAddressAnnotationTemplate, podNet.ProviderName)]
 	if macStr != "" {
 		if _, err := net.ParseMAC(macStr); err != nil {
@@ -1203,9 +1225,9 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 		pod.Annotations[fmt.Sprintf(util.IpPoolAnnotationTemplate, podNet.ProviderName)] == "" {
 		var skippedAddrs []string
 		for {
-			nicName := ovs.PodNameToPortName(podName, pod.Namespace, podNet.ProviderName)
+			portName := ovs.PodNameToPortName(podName, pod.Namespace, podNet.ProviderName)
 
-			ipv4, ipv6, mac, err := c.ipam.GetRandomAddress(key, nicName, macStr, podNet.Subnet.Name, skippedAddrs, !podNet.AllowLiveMigration)
+			ipv4, ipv6, mac, err := c.ipam.GetRandomAddress(key, portName, macStr, podNet.Subnet.Name, skippedAddrs, !podNet.AllowLiveMigration)
 			if err != nil {
 				return "", "", "", podNet.Subnet, err
 			}
@@ -1226,7 +1248,7 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 		}
 	}
 
-	nicName := ovs.PodNameToPortName(podName, pod.Namespace, podNet.ProviderName)
+	portName := ovs.PodNameToPortName(podName, pod.Namespace, podNet.ProviderName)
 
 	// The static ip can be assigned from any subnet after ns supports multi subnets
 	nsNets, _ := c.getNsAvailableSubnets(pod, podNet)
@@ -1238,7 +1260,7 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 		ipStr := pod.Annotations[fmt.Sprintf(util.IpAddressAnnotationTemplate, podNet.ProviderName)]
 
 		for _, net := range nsNets {
-			v4IP, v6IP, mac, err = c.acquireStaticAddress(key, nicName, ipStr, macStr, net.Subnet.Name, net.AllowLiveMigration)
+			v4IP, v6IP, mac, err = c.acquireStaticAddress(key, portName, ipStr, macStr, net.Subnet.Name, net.AllowLiveMigration)
 			if err == nil {
 				return v4IP, v6IP, mac, net.Subnet, nil
 			}
@@ -1252,7 +1274,7 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 		ipPool[i] = strings.TrimSpace(ip)
 	}
 
-	if ok, _ := isStatefulSetPod(pod); !ok {
+	if !isStsPod {
 		for _, net := range nsNets {
 			for _, staticIPs := range ipPool {
 				for _, staticIP := range strings.Split(staticIPs, ",") {
@@ -1262,7 +1284,7 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 					}
 				}
 
-				v4IP, v6IP, mac, err = c.acquireStaticAddress(key, nicName, staticIPs, macStr, net.Subnet.Name, net.AllowLiveMigration)
+				v4IP, v6IP, mac, err = c.acquireStaticAddress(key, portName, staticIPs, macStr, net.Subnet.Name, net.AllowLiveMigration)
 				if err == nil {
 					return v4IP, v6IP, mac, net.Subnet, nil
 				}
@@ -1275,7 +1297,7 @@ func (c *Controller) acquireAddress(pod *v1.Pod, podNet *kubeovnNet) (string, st
 		index, _ := strconv.Atoi(numStr)
 		if index < len(ipPool) {
 			for _, net := range nsNets {
-				v4IP, v6IP, mac, err = c.acquireStaticAddress(key, nicName, ipPool[index], macStr, net.Subnet.Name, net.AllowLiveMigration)
+				v4IP, v6IP, mac, err = c.acquireStaticAddress(key, portName, ipPool[index], macStr, net.Subnet.Name, net.AllowLiveMigration)
 				if err == nil {
 					return v4IP, v6IP, mac, net.Subnet, nil
 				}
@@ -1344,6 +1366,7 @@ func appendCheckPodToDel(c *Controller, pod *v1.Pod, ownerRefName, ownerRefKind 
 
 	// check if subnet exist in OwnerReference
 	var ownerRefSubnetExist bool
+	var ownerRefSubnet string
 	switch ownerRefKind {
 	case "StatefulSet":
 		ss, err := c.config.KubeClient.AppsV1().StatefulSets(pod.Namespace).Get(context.Background(), ownerRefName, metav1.GetOptions{})
@@ -1356,6 +1379,7 @@ func appendCheckPodToDel(c *Controller, pod *v1.Pod, ownerRefName, ownerRefKind 
 		}
 		if ss.Spec.Template.ObjectMeta.Annotations[util.LogicalSwitchAnnotation] != "" {
 			ownerRefSubnetExist = true
+			ownerRefSubnet = ss.Spec.Template.ObjectMeta.Annotations[util.LogicalSwitchAnnotation]
 		}
 
 	case util.VmInstance:
@@ -1372,6 +1396,7 @@ func appendCheckPodToDel(c *Controller, pod *v1.Pod, ownerRefName, ownerRefKind 
 			vm.Spec.Template.ObjectMeta.Annotations != nil &&
 			vm.Spec.Template.ObjectMeta.Annotations[util.LogicalSwitchAnnotation] != "" {
 			ownerRefSubnetExist = true
+			ownerRefSubnet = vm.Spec.Template.ObjectMeta.Annotations[util.LogicalSwitchAnnotation]
 		}
 	}
 
@@ -1391,6 +1416,11 @@ func appendCheckPodToDel(c *Controller, pod *v1.Pod, ownerRefName, ownerRefKind 
 	}
 	if podSubnet != nil && !util.CIDRContainIP(podSubnet.Spec.CIDRBlock, pod.Annotations[util.IpAddressAnnotation]) {
 		klog.Infof("pod's ip %s is not in the range of subnet %s, delete pod", pod.Annotations[util.IpAddressAnnotation], podSubnet.Name)
+		return true, nil
+	}
+	// subnet of ownerReference(sts/vm) has been changed, it needs to handle delete pod and create port on the new logical switch
+	if podSubnet != nil && ownerRefSubnet != "" && podSubnet.Name != ownerRefSubnet {
+		klog.Infof("Subnet of owner %s has been changed from %s to %s, delete pod %s/%s", ownerRefName, podSubnet.Name, ownerRefSubnet, pod.Namespace, pod.Name)
 		return true, nil
 	}