+++ pkg/controller/service.go
@@ -1,11 +1,14 @@
 package controller
 
 import (
+	"context"
 	"fmt"
 	"strings"
+	"time"
 
 	v1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/errors"
+	k8serrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/tools/cache"
@@ -19,6 +22,7 @@ type vpcService struct {
 	Vip      string
 	Vpc      string
 	Protocol v1.Protocol
+	Svc      *v1.Service
 }
 
 func (c *Controller) enqueueAddService(obj interface{}) {
@@ -32,7 +36,6 @@ func (c *Controller) enqueueAddService(obj interface{}) {
 		return
 	}
 	svc := obj.(*v1.Service)
-	klog.V(3).Infof("enqueue add service %s", key)
 
 	if c.config.EnableNP {
 		var netpols []string
@@ -45,6 +48,11 @@ func (c *Controller) enqueueAddService(obj interface{}) {
 			c.updateNpQueue.Add(np)
 		}
 	}
+
+	if c.config.EnableLbSvc {
+		klog.V(3).Infof("enqueue add service %s", key)
+		c.addServiceQueue.Add(key)
+	}
 }
 
 func (c *Controller) enqueueDeleteService(obj interface{}) {
@@ -54,7 +62,9 @@ func (c *Controller) enqueueDeleteService(obj interface{}) {
 	svc := obj.(*v1.Service)
 	//klog.V(3).Infof("enqueue delete service %s/%s", svc.Namespace, svc.Name)
 	klog.Infof("enqueue delete service %s/%s", svc.Namespace, svc.Name)
-	if svc.Spec.ClusterIP != v1.ClusterIPNone && svc.Spec.ClusterIP != "" {
+
+	vip, ok := svc.Annotations[util.SwitchLBRuleVipsAnnotation]
+	if ok || svc.Spec.ClusterIP != v1.ClusterIPNone && svc.Spec.ClusterIP != "" {
 
 		if c.config.EnableNP {
 			var netpols []string
@@ -68,11 +78,18 @@ func (c *Controller) enqueueDeleteService(obj interface{}) {
 				c.updateNpQueue.Add(np)
 			}
 		}
+
+		ip := svc.Spec.ClusterIP
+		if ok {
+			ip = vip
+		}
+
 		for _, port := range svc.Spec.Ports {
 			vpcSvc := &vpcService{
-				Vip:      fmt.Sprintf("%s:%d", svc.Spec.ClusterIP, port.Port),
+				Vip:      fmt.Sprintf("%s:%d", ip, port.Port),
 				Protocol: port.Protocol,
 				Vpc:      svc.Annotations[util.VpcAnnotation],
+				Svc:      svc,
 			}
 			klog.Infof("delete vpc service %v", vpcSvc)
 			c.deleteServiceQueue.Add(vpcSvc)
@@ -100,6 +117,11 @@ func (c *Controller) enqueueUpdateService(old, new interface{}) {
 	c.updateServiceQueue.Add(key)
 }
 
+func (c *Controller) runAddServiceWorker() {
+	for c.processNextAddServiceWorkItem() {
+	}
+}
+
 func (c *Controller) runDeleteServiceWorker() {
 	for c.processNextDeleteServiceWorkItem() {
 	}
@@ -110,9 +132,38 @@ func (c *Controller) runUpdateServiceWorker() {
 	}
 }
 
+func (c *Controller) processNextAddServiceWorkItem() bool {
+	obj, shutdown := c.addServiceQueue.Get()
+	if shutdown {
+		return false
+	}
+
+	err := func(obj interface{}) error {
+		defer c.addServiceQueue.Done(obj)
+		var key string
+		var ok bool
+		if key, ok = obj.(string); !ok {
+			c.addServiceQueue.Forget(obj)
+			utilruntime.HandleError(fmt.Errorf("expected string in workqueue but got %#v", obj))
+			return nil
+		}
+		if err := c.handleAddService(key); err != nil {
+			c.addServiceQueue.AddRateLimited(key)
+			return fmt.Errorf("error syncing '%s': %s, requeuing", key, err.Error())
+		}
+		c.addServiceQueue.Forget(obj)
+		return nil
+	}(obj)
+
+	if err != nil {
+		utilruntime.HandleError(err)
+		return true
+	}
+	return true
+}
+
 func (c *Controller) processNextDeleteServiceWorkItem() bool {
 	obj, shutdown := c.deleteServiceQueue.Get()
-
 	if shutdown {
 		return false
 	}
@@ -207,6 +258,13 @@ func (c *Controller) handleDeleteService(service *vpcService) error {
 		}
 	}
 
+	if service.Svc.Spec.Type == v1.ServiceTypeLoadBalancer && c.config.EnableLbSvc {
+		if err := c.deleteLbSvc(service.Svc); err != nil {
+			klog.Errorf("failed to delete service %s, %v", service.Svc.Name, err)
+			return err
+		}
+	}
+
 	return nil
 }
 
@@ -219,15 +277,20 @@ func (c *Controller) handleUpdateService(key string) error {
 	klog.Infof("update svc %s/%s", namespace, name)
 	svc, err := c.servicesLister.Services(namespace).Get(name)
 	if err != nil {
-		if errors.IsNotFound(err) {
+		if k8serrors.IsNotFound(err) {
 			return nil
 		}
 		return err
 	}
 
-	ip := svc.Spec.ClusterIP
-	if ip == "" || ip == v1.ClusterIPNone {
-		return nil
+	ip := ""
+	if vip, ok := svc.Annotations[util.SwitchLBRuleVipsAnnotation]; ok {
+		ip = vip
+	} else {
+		ip = svc.Spec.ClusterIP
+		if ip == "" || ip == v1.ClusterIPNone {
+			return nil
+		}
 	}
 
 	tcpVips := []string{}
@@ -335,7 +398,6 @@ func (c *Controller) handleUpdateService(key string) error {
 	return nil
 }
 
-// The type of vips is map, which format is like [fd00:10:96::11c9]:10665:[fc00:f853:ccd:e793::2]:10665,[fc00:f853:ccd:e793::3]:10665
 // Parse key of map, [fd00:10:96::11c9]:10665 for example
 func parseVipAddr(vipStr string) string {
 	vip := strings.Split(vipStr, ":")[0]
@@ -343,4 +405,90 @@ func parseVipAddr(vipStr string) string {
 		vip = strings.Trim(strings.Split(vipStr, "]")[0], "[]")
 	}
 	return vip
+}
+
+func (c *Controller) handleAddService(key string) error {
+	namespace, name, err := cache.SplitMetaNamespaceKey(key)
+	if err != nil {
+		utilruntime.HandleError(fmt.Errorf("invalid resource key: %s", key))
+		return nil
+	}
+
+	svc, err := c.servicesLister.Services(namespace).Get(name)
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			return nil
+		}
+		return err
+	}
+	if svc.Spec.Type != v1.ServiceTypeLoadBalancer || !c.config.EnableLbSvc {
+		return nil
+	}
+	klog.Infof("add svc %s/%s", namespace, name)
+
+	if err = c.validateSvc(svc); err != nil {
+		klog.Errorf("failed to validate lb svc, %v", err)
+		return err
+	}
+
+	if err = c.checkAttachNetwork(svc); err != nil {
+		klog.Errorf("failed to check attachment network, %v", err)
+		return err
+	}
+
+	if err = c.createLbSvcPod(svc); err != nil {
+		klog.Errorf("failed to create lb svc pod, %v", err)
+		return err
+	}
+
+	var pod *v1.Pod
+	for {
+		pod, err = c.getLbSvcPod(name, namespace)
+		if err != nil {
+			klog.Errorf("wait lb svc pod to running, %v", err)
+			time.Sleep(1 * time.Second)
+		}
+		if pod != nil {
+			break
+		}
+
+		// It's important here to check existing of svc, used to break the loop.
+		_, err = c.servicesLister.Services(namespace).Get(name)
+		if err != nil {
+			if k8serrors.IsNotFound(err) {
+				return nil
+			}
+			return err
+		}
+	}
+
+	loadBalancerIP, err := c.getPodAttachIP(pod, svc)
+	if err != nil {
+		klog.Errorf("failed to get loadBalancerIP: %v", err)
+		return err
+	}
+
+	newSvc, err := c.servicesLister.Services(namespace).Get(name)
+	if err != nil {
+		if k8serrors.IsNotFound(err) {
+			return nil
+		}
+		return err
+	}
+	var ingress v1.LoadBalancerIngress
+	ingress.IP = loadBalancerIP
+	newSvc.Status.LoadBalancer.Ingress = []v1.LoadBalancerIngress{ingress}
+
+	var updateSvc *v1.Service
+	if updateSvc, err = c.config.KubeClient.CoreV1().Services(namespace).UpdateStatus(context.Background(), newSvc, metav1.UpdateOptions{}); err != nil {
+		klog.Errorf("update service %s/%s status failed: %v", namespace, name, err)
+		return err
+	}
+
+	if err := c.updatePodAttachNets(pod, updateSvc); err != nil {
+		klog.Errorf("update service %s/%s attachment network failed: %v", namespace, name, err)
+		return err
+	}
+
+	return nil
 }